+++ persistence/types/mocks/mocks.go
@@ -1,0 +1,3 @@
+package mock_kvstore
+
+// This file is in place to declare the package for dynamically generated mocks