+++ consensus/doc/CHANGELOG.md
@@ -7,15 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.27] - 2023-02-09
+## [0.0.0.28] - 2023-02-10
 
 - Modules embed `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
 - Updated modules `Create` to accept generic options
 - `resetToGenesis` clears the utility mempool as well
 - Publishing `ConsensusNewHeightEvent` on new height
 
-## [0.0.0.26] - 2023-02-06
+## [0.0.0.27] - 2023-02-09
+- Add `state_sync` submodule, with `state_sync` struct
+- Implement state sync server to advertise blocks and metadata
+- Create new `state_sync_handler.go` source file that handles `StateSyncMessage`s sent to the `Consensus` module
+- Add two new tests in `state_sync_test.go`:`TestStateSyncServerGetMetaDataReq` and `TestStateSyncServerGetBlock`
+- Update `TestHotstuff4Nodes1BlockHappyPath` test to also retrieve the committed block
 
+## [0.0.0.26] - 2023-02-06
 - Address legacy linter errors from `golangci-lint`
 
 ## [0.0.0.25] - 2023-02-04