+++ logger/docs/CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.6] - 2023-02-09
+## [0.0.0.7] - 2023-02-10
 
 - Module embeds `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
+
+## [0.0.0.6] - 2023-02-09
+
+- `loggerModule` type-checking for `modules.Module`
 
 ## [0.0.0.5] - 2023-02-07