+++ p2p/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.25] - 2023-02-09
+## [0.0.0.26] - 2023-02-10
 
 - Modules embed `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
 - Deprecated `debugAddressBookProvider`
@@ -20,6 +20,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added basic `bootstrap` nodes support
 - Reacting to `ConsensusNewHeightEventType` adn `StateMachineTransitionEventType` to update the address book and current height and determine if a bootstrap is needed
 - Updated tests
+
+## [0.0.0.25] - 2023-02-09
+
+- Updated logging initialization and passing to the network component instead of using the global logger
+- Fixed incorrect use of `bus.GetLoggerModule()` in `stdnetwork.go` since it's never initialized when running the debug CLI
 
 ## [0.0.0.24] - 2023-02-06