+++ runtime/docs/CHANGELOG.md
@@ -7,12 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.16] - 2023-02-09
+## [0.0.0.17] - 2023-02-10
 
 - Added `bootstrap_nodes_csv` in `P2PConfig` to allow for a comma separated list of bootstrap nodes
 - Introduced `modules.ModulesRegistry` for better separation of concerns
 - Added `StateMachineModule` accessors
 - `Manager` embeds `modules.BaseIntegratableModule` for DRYness
+
+## [0.0.0.16] - 2023-02-09
+
+- Update runtime consensus config with bool server mode variable
+- Update manager test
 
 ## [0.0.0.15] - 2023-02-07